add handleArrowButtons tests
